+++ changelog.md
@@ -3,6 +3,8 @@
 ## Next
 - Blocks are now placed adjacent to the hovered block face
 - Block texels are now larger
+- Block textures are now applied to each face individually
+- Fixed orientation of block textures
 
 ## 0.0.18_1
 *2024-07-25 16:45*